requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dateutil==2.8.2
Pillow==10.2.0
Werkzeug==3.0.1
